+++ src/controllers/skill/implementation/skill.controller.ts
@@ -3,6 +3,7 @@ import { ISkillController } from "../interfaces/skill.controller.interface";
 import { ISkillServivce } from "../../../services/skills/interfaces/skills.services.interfaces";
 import { ApiResponse } from "../../../utils/apiResponse.utils";
 import { ISkill } from "../../../models/skill/skill.interface";
+import { ValidationError } from "../../../errors-classes/validation.error";
 
 export class SkillController implements ISkillController {
     constructor(private _skillService: ISkillServivce) {}
@@ -20,8 +21,6 @@ export class SkillController implements ISkillController {
     }
 
     async getPaginatedSkills(req: Request, res: Response, next: NextFunction): Promise<Response | void> {
-
-        
         try {
             const page = Number(req.query.page) || 1;
             const limit = Number(req.query.limit) || 10;
@@ -40,6 +39,9 @@ export class SkillController implements ISkillController {
         try {
             const skillId = req.params.id;
 
+            if (!skillId || typeof skillId !== "string") {
+                throw new ValidationError("Invalid skillId");
+            }
             const skill = await this._skillService.getSkillById(skillId);
             return ApiResponse.success<ISkill | null>(res, "Skill fetched successfully", skill);
         } catch (error) {
@@ -49,7 +51,12 @@ export class SkillController implements ISkillController {
 
     async updateSkill(req: Request, res: Response, next: NextFunction): Promise<Response | void> {
         try {
-            const updatedSkill = await this._skillService.updateSkill(req.params.id, req.body);
+            const id = req.params.id;
+            if (!id || typeof id !== "string") {
+                throw new ValidationError("Invalid id");
+            }
+
+            const updatedSkill = await this._skillService.updateSkill(id, req.body);
 
             return ApiResponse.success(res, "Skill updated", updatedSkill);
         } catch (error) {
@@ -60,7 +67,9 @@ export class SkillController implements ISkillController {
     async blockOrUnblockSkill(req: Request, res: Response, next: NextFunction): Promise<Response | void> {
         try {
             const skillId = req.params.id;
-
+            if (!skillId || typeof skillId !== "string") {
+                throw new ValidationError("Invalid skillId");
+            }
             const skill = await this._skillService.blockOrUnblockSkill(skillId);
 
             return ApiResponse.success(res, "Skill block or unblock done", skill);
@@ -72,7 +81,9 @@ export class SkillController implements ISkillController {
     async deleteSkill(req: Request, res: Response, next: NextFunction): Promise<Response | void> {
         try {
             const skillId = req.params.id;
-
+            if (!skillId || typeof skillId !== "string") {
+                throw new ValidationError("Invalid skillId");
+            }
             await this._skillService.deleteSkill(skillId);
 
             return ApiResponse.success(res, "Skill deleted successfully");
